File_ops.py Created
- Moved file operations from algorithms.py to file_ops.py
