+++ morphengine.py
@@ -29,7 +29,7 @@ import bpy
 
 import mathutils
 
-from . import algorithms, utils, proxyengine
+from . import algorithms, utils, proxyengine, file_ops
 
 import time, json
 import operator
@@ -44,7 +44,7 @@ class MorphingEngine:
 
     def __init__(self, obj_name, character_config):
         time1 = time.time()
-        data_path = algorithms.get_data_path()
+        data_path = file_ops.get_data_path()
         self.final_form = []
         self.cache_form = []
         self.obj_name = obj_name
@@ -123,7 +123,7 @@ class MorphingEngine:
         self.proportion_index = None
 
         self.init_final_form()
-        self.base_form = algorithms.load_vertices_database(self.vertices_path)
+        self.base_form = file_ops.load_vertices_database(self.vertices_path)
 
         self.load_morphs_database(self.shared_morph_data_path)
         self.load_morphs_database(self.morph_data_path)
@@ -158,7 +158,7 @@ class MorphingEngine:
         return None
 
     def error_msg(self, path):
-        logger.warning("Database file not found: {0}".format(algorithms.simple_path(path)))
+        logger.warning("Database file not found: {0}".format(file_ops.simple_path(path)))
 
     def reset(self, update=True):
         for i in range(len(self.base_form)):
@@ -169,7 +169,7 @@ class MorphingEngine:
             self.update(update_all_verts=True)
 
     def load_measures_database(self, measures_path):
-        m_database = algorithms.load_json_data(measures_path,"Measures data")
+        m_database = file_ops.load_json_data(measures_path,"Measures data")
         if m_database:
             self.measures_data = m_database["measures"]
             self.measures_relat_data = m_database["relations"]
@@ -177,12 +177,12 @@ class MorphingEngine:
             self.body_height_Z_parts = m_database["body_height_Z_parts"]
 
     def load_bboxes_database(self, bounding_box_path):
-        self.bbox_data = algorithms.load_json_data(bounding_box_path,"Bounding box data")
+        self.bbox_data = file_ops.load_json_data(bounding_box_path,"Bounding box data")
 
 
     def load_morphs_database(self, morph_data_path):
         time1 = time.time()
-        m_data = algorithms.load_json_data(morph_data_path,"Morph data") # calls algorithms.py
+        m_data = file_ops.load_json_data(morph_data_path,"Morph data")
         if m_data:
             for morph_name, deltas in m_data.items():
                 morph_deltas = []
@@ -197,7 +197,7 @@ class MorphingEngine:
                 self.morph_data[morph_name] = morph_deltas
                 self.morph_values[morph_name] = 0.0
                 self.morph_modified_verts[morph_name] = modified_verts
-            logger.info("Morph database {0} loaded in {1} secs".format(algorithms.simple_path(morph_data_path),time.time()-time1))
+            logger.info("Morph database {0} loaded in {1} secs".format(file_ops.simple_path(morph_data_path),time.time()-time1))
             logger.info("Now local morph data contains {0} elements".format(len(self.morph_data)))
 
 
@@ -253,7 +253,7 @@ class MorphingEngine:
 
 
     def compare_file_proportions(self,filepath):
-        char_data = algorithms.load_json_data(filepath,"Proportions data")
+        char_data = file_ops.load_json_data(filepath,"Proportions data")
         if "proportion_index" in char_data:
             v1 = mathutils.Vector(self.proportion_index)
             v2 = mathutils.Vector(char_data["proportion_index"])
@@ -261,7 +261,7 @@ class MorphingEngine:
 
             return (delta_v.length,filepath)
         else:
-            logger.info("File {0} does not contain proportions".format(algorithms.simple_path(filepath)))
+            logger.info("File {0} does not contain proportions".format(file_ops.simple_path(filepath)))
 
 
     def compare_data_proportions(self):
